consolidates mobile app
