+++ app/mobile/Main.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Router, Reducer, ActionConst } from 'react-native-router-flux'
-import routes from './routes';
+import views from './views';
 
 const reducer = params => {
   const defaultReducer = new Reducer(params);
@@ -10,14 +10,12 @@ const reducer = params => {
   };
 };
 
-class Main extends React.Component {
+export default class Main extends React.Component {
   render() {
     return (
       <Router createReducer={reducer} tintColor='red'>
-        {routes()}
+        {views()}
       </Router>
     );
   }
 }
-
-export default Main;
