+++ index.android.js
@@ -1,4 +1,13 @@
+import React, { Component } from 'react';
 import { AppRegistry } from 'react-native';
-import Main from './app/Main.js';
+import Main from './app/mobile/Main.js';
 
-AppRegistry.registerComponent('aandr_wedding_app', () => Main);
+class aandr_wedding_app extends Component {
+  render() {
+    return (
+      <Main />
+    );
+  }
+}
+
+AppRegistry.registerComponent('aandr_wedding_app', () => aandr_wedding_app);
